+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 ### Version 2.0.9
+- FIX: Show an error in impact and pyramid view if no ingredients
 
 ### Version 2.0.8
 - ADD: Integration with PV Online